* Initial xsh serial test

* Clean branch for formal release.

* Remove old platforms

* Remove old archs

* Fix aarch64 platformvars

* Change aarch64 to armv8

* added fpu and abi

* WIP - FPU. Add double test

* start.S changes

* WIP framebuffer stuck at mbox read

* WIP framebuffer+usbkbd in a working state

* changed color hex format

* framebuffer minishell WIP

* Framebuffer driver complete

* Update ready for multicore. Add CPUID for ps cmd

* Add mutexes

* WIP test multicore kprintf

* cleaned up swap files, and added to gitignore

* WIP mutex table

* WIP 3-cores breaks mutex

* Merge homework MMU changes

* Tempfix mmu, test suite compiling

* added debug code

* Doxy branch

* WIP mmu lan7800 conflict discovered

* kill.c changes. shell banner fixed

* WIP usb dma cache issue

* WIP flush cache

* Add cache arch info

* Test suite lan disabled

* lan7800 dox formatting

* Frame Buffer dox formatting

* WIP changed cache flush increment

* WIP initial priority/aging test

* fixed arm-rpi3 dox formatting

* WIP cache flush behavior analysis

* added priority scheduling test

* WIP add cache area functions

* WIP better state - device close to init

* WIP fixing dma for config xfer

* priority scheduling

* added Doxygen.PROJECT_NUMBER

* Rade did some stuff

* it works now

* added another priority scheduling test case

* Minor cleaning, WIP bLength

* Switch to unaligned accessing

* Kprintf fixes execution temporarily

* WIP mutex lock conflict

* WIP endpoints invalid for LAN

* WIP LAN device config desc fails

* fixed some test cases (getcpuid fixes a lot)

* added multicore to schedule test

* WIP Add mmu page function

* Semaphore2 test has multicore now

* semaphore2 works i guess

* WIP printf testing

* attempted to create non-cacheable buffer space for dma transfers to occur

* Add LAN docs to doxy

* WIP fixing doxy mainpage to readme.md

* Major doc updates

* All TODODOC fixed

* printf fixup

* Merge testsuite changes into doxy branch

* Post-merge testsuite branch

* Add shell updates

* fixed multiple semaphore test

* usb devices attach.. ethernet device is not getting mac address set..

* Fix shell hang due to improper invalidation

* usb working now with 2d array abstraction.. yuck.

* LAN+mcore working state. Cleanup to do

* minor cleaning updates

* removed 2d array abstraction

* framebuffer mailbox.. attempting to use dma_buf

* Fix framebuffer. WIP cleanup

* cleaning up.. may need more cleaning.. not sure

* WIP framebuffer cleanup

* WIP frame buffer escape states

* Add color parsing to fbPutc

* Add shell clear to fbPutc parser

* Atomic operations for rpi3

+++ shell/xsh_kexec.c
@@ -178,7 +178,7 @@ static void kexec_from_network(int netdev)
            data.bootfile, str_ip);
     kernel = (void*)tftpGetIntoBuffer(data.bootfile, &nif->ip,
                                       &data.next_server, &size);
-    printf("kernel=0x%08X, *kernel=0x%08X\n", kernel, *(uint*)kernel);
+    printf("kernel=0x%08X, *kernel=0x%08X\n", (uint)kernel, *(uint *)kernel);
 
     if (SYSERR == (int)kernel)
     {
